feat(ui): refactor intro component
