+++ configs/prod.yaml
@@ -4,4 +4,3 @@ security:
     enabled: true
 features:
   cut: []
-  customContent: []
