+++ src/ui/components/CustomContent/Intro/Intro.scss
