+++ src/ui/components/CustomContent/Intro/Intro.types.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from "react";
+
+interface SlideItem {
+  image: string;
+  description?: string;
+  title?: string;
+  lottie?: ReactNode;
+}
+
+export type { SlideItem };
