+++ src/ui/components/CustomContent/Intro/index.ts
@@ -1,0 +1,1 @@
+export * from "./Intro";
