--- src/ui/pages/Onboarding/components/Slides/index.ts
@@ -1,1 +1,0 @@
-export * from "./Slides";
